+++ info/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2025-08-25
+
+### Modified
+
+- Rewrite installed application enumeration helper to collect more data, and especially the application version, where possible.
+- Split the installed application check into two separate checks, one for Microsoft products, and the other for third-party apps.
+
 ## 2025-08-19
 
 ### Added